Feature/lens distortion (#69)
* Lens Distortion now added as a pass for both semantic and instance segmentation by way of a LensDistortion shader.  This can be enabled either through enabling post-processing in URP or HDRP respectively or by utilizing the override although this approach is internal only currently but allows for the manipulation of the shader without having to have access to the global volume which in some test scenarios is not available.

+++ com.unity.perception/Runtime/GroundTruth/Resources/InstanceSegmentation.shader
@@ -41,7 +41,12 @@
 
             fixed4 frag (v2f i) : SV_Target
             {
-                return float4(UnpackUIntToFloat((uint)_SegmentationId, 0, 8), UnpackUIntToFloat(_SegmentationId, 8, 8), UnpackUIntToFloat(_SegmentationId, 16, 8), UnpackUIntToFloat(_SegmentationId, 24, 8));
+                return fixed4(
+                    UnpackUIntToFloat((uint)_SegmentationId,  0, 8),
+                    UnpackUIntToFloat((uint)_SegmentationId,  8, 8),
+                    UnpackUIntToFloat((uint)_SegmentationId, 16, 8),
+                    UnpackUIntToFloat((uint)_SegmentationId, 24, 8)
+                );
             }
             ENDCG
         }
